+++ NEWS.md
@@ -4,6 +4,9 @@
 * New `rscript()` function and  `rscript_process` class to execute
   R scripts via `Rscript` (#40, #81).
 
+* Library paths are now correctly set set up for `system()` (and similar)
+  calls from the callr subprocesses (#83, #84).
+
 # callr 3.0.0
 
 * New `r_session` class: a background R session you can send commands to